// Harken-v2-release2025.08/Harken-v2-release2025.08/packages/backend/src/utils/enums/MessageEnum.ts
export enum ClientEnum {
	CLIENT_SAVE_SUCCESS = 'Client Saved Successfully',
	CLIENT_SAVE_FAIL = 'Client Save fail',
	CLIENTS_LIST = 'Clients list',
	CLIENTS_NOT_FOUND = 'Clients not found',
	CLIENT_UPDATE_FAIL = 'Clients not update',
	PERMISSION_DENIED = 'Do not have permission for this user',
	CLIENT_DELETE = 'Client deleted successfully',
	DELETE_CLIENT_FAIL = 'Client not deleted ',
	CLIENT_DATA = 'Client data',
	NOT_FOUND = 'Client not found',
}